deduplicate docker test helper functions
code was becoming duplicated too much, moved to TestHelper.scala
